requirements.txt
spotipy==2.12.0
Flask==1.0.2
gunicorn==19.4.5
itsdangerous==0.24
Jinja2==2.10.3
MarkupSafe==0.23
requests==2.9.1
Werkzeug==0.16.0
wheel==0.24.0
yolk==0.4.3
Flask-WTF==0.13
urllib3==1.25.9
psycopg2-binary==2.8.6
